requirements.txt
pytesseract
mysql-connector-python
opencv-python
matplotlib
tkcalendar
pyserial